requirements.txt
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
APScheduler==3.10.4
pytz==2023.3